refactor: infrastructure/persistence/ を infrastructure/repository/ に統合
persistence/ と repository/ が並立していたパッケージ構造を
repository/ に一本化し、Spring Data JDBC 関連の実装を集約した。

Closes #105
