+++ backend/src/main/java/com/youmorry/expensetracker/infrastructure/repository/JdbcConfiguration.java
@@ -1,7 +1,7 @@
-package com.youmorry.expensetracker.infrastructure.persistence;
+package com.youmorry.expensetracker.infrastructure.repository;
 
-import com.youmorry.expensetracker.infrastructure.persistence.converter.IdConverters;
-import com.youmorry.expensetracker.infrastructure.persistence.converter.ValueObjectConverters;
+import com.youmorry.expensetracker.infrastructure.repository.converter.IdConverters;
+import com.youmorry.expensetracker.infrastructure.repository.converter.ValueObjectConverters;
 import java.time.Clock;
 import java.time.Instant;
 import java.util.List;
